+++ types/gmp.interface.ts
@@ -1,0 +1,10 @@
+interface IPOData {
+    image: string;
+    title: string;
+    offer_date: string;
+    price: string;
+    gmp: string;
+    gain: string;
+    label: string;
+    url: string;
+}
